chore: fix CI — split SelectListItem into its own file, expression body
- Move SelectListItem + variant helpers out of ListItem.kt into new
  SelectListItem.kt to clear detekt TooManyFunctions (ListItem.kt was at
  the 15-function threshold).
- Change private ListItem composable helpers to internal so the extracted
  PlainSelectListItem can still delegate to them.
- Convert Set<Int>.toggle in SelectListItemDisplay.kt and SelectListItemType.role
  getter to expression body (ktlint function-expression-body rule).
